// src/app/[lang]/flights/types/flight.ts
export interface FlightClientVO {
  code: string;
  airlineName: string;
  departureCity: string;
  departureAirport: string;
  departureIATA: string;
  arrivalCity: string;
  arrivalAirport: string;
  arrivalIATA: string;
  departureTime: string;
  dateTimeArrival: string;
  price: number;
  seatsLeft: number;
}